10/10 2

+++ src/main/java/com/cnpm/demo/model/Model/Employee.java
@@ -1,6 +1,11 @@
 package com.cnpm.demo.model.Model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "employee")
@@ -36,7 +41,7 @@ public class Employee {
 
     // Getters và Setters
 
-    public int getIdEmployee() {
+    public long getIdEmployee() {
         return idEmployee;
     }
